Finish application logic, see known issues below
There is a problem with unlockItem function, it crashes upon calling
findViewById despite correct Id number. I'm not sure what to do with it
yet, other than that app is working correctly, but no effort was made
to make it visually appealing. You pay with clicks to unlock better
buttons, you can hold the upgrade button to check it's price.